Use Case Number and Case Name to filter for emails (#277)
Use Case Number and Case Name to filter for emails
Replace retry with PollyRetry

This change should handle the concurrent tests running to make sure the email assertion is against the current booking and not against a contact email and template id. If 2 bookings are made at the same time, we should make sure we're not passing on the second booking passing the first booking assertions.